WIP - get around houses

+++ src/main/kotlin/playwithyou/lucca/domain/entity/Coordinates.kt
@@ -1,5 +1,5 @@
 package playwithyou.lucca.domain.entity
 
-data class Coordinates(val latitude: Float, val longitude: Float) {
+data class Coordinates(val latitude: Double, val longitude: Double) {
 
 }
